MoCo_CL/requirements.txt
imageio==2.34.2+computecanada
keyring==23.5.0
kiwisolver==1.4.5+computecanada
numpy==1.26.4+computecanada
packaging==21.3
pandas==2.2.1+computecanada
Pillow==10.0.0+computecanada
scikit-image==0.24.0+computecanada
scipy==1.10.1+computecanada
six==1.16.0
tensorboardX==2.6.2.2
timm==0.9.2
torch==2.1.1+computecanada
torchmetrics==1.4.0.post0
torchsummary==1.5.1
torchvision==0.16.1+computecanada
tqdm==4.64.1
yacs==0.1.8
yarl==1.9.4
zipp==1.0.0
matplotlib==3.9.0+computecanada